Almost there.

+++ scripts/unitManager.py
@@ -164,7 +164,7 @@ class UnitManager(object):
                         storages[agent.agentName] = thisStorage
         return storages
 
-    def removeInstance(self, arg):
+    def removeInstance(self, arg, soft=False):
         '''
         Process the destruction of a unit
         :param fifeID: FifeID or instace or agent of the destroyed unit
@@ -181,7 +181,8 @@ class UnitManager(object):
             self.fife2Agent.__delitem__(fifeID)
             agent.instance.removeActionListener(agent)
 
-            self.agentLayer.deleteInstance(agent.instance)
+            if not soft:
+                self.agentLayer.deleteInstance(agent.instance)
             agent.instance = None
 
             if fifeID in self.dummyIDs.keys():
